Reorganise sparse indexing tests into a separate file
